+++ src/test/java/carracing/domain/CarTest.java
@@ -1,7 +1,8 @@
-package carracing;
+package carracing.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import carracing.domain.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
